+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Add ability to toggle the powerups to the settings page
 * Clicking on the same creature/egg row will hide the actions row
 * Added Creature select action
+* Fix certain s16 sprites being too big on first show (object-fit bug)
 
 ## 0.1.1 (2018-03-12)